+++ src/core/async/modules/events/index.ts
@@ -12,29 +12,39 @@
  */
 
 import SyncPromise from 'core/promise/sync';
-import Super, { isEvent } from 'core/async/modules/timers';
+
+import Super, {
+
+	IdObject,
+	ProxyCb,
+
+	StrictClearOptionsId
+
+} from 'core/async/modules/timers';
+
+import { isEvent } from 'core/async/modules/events/helpers';
 
 import type {
 
+	Event,
+	EventId,
+	EventEmitterLikeP,
+
 	AsyncOnOptions,
 	AsyncOnceOptions,
-	AsyncPromisifyOnceOptions,
+	AsyncPromisifyOnceOptions
 
-	EventId,
-	ClearOptionsId,
-
-	ProxyCb,
-	Event,
-	EventEmitterLikeP
-
-} from 'core/async/interface';
+} from 'core/async/modules/events/interface';
 
 export * from 'core/async/modules/timers';
+export * from 'core/async/modules/events/helpers';
+export * from 'core/async/modules/events/interface';
 
 export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	/**
 	 * Attaches an event listener from the specified event emitter.
 	 * If the emitter is a function, it is interpreted as the function to attach events.
+	 * Notice, if you don't provide a group for the operation, it will be taken from the event name.
 	 *
 	 * @param emitter - event emitter
 	 * @param events - event or list of events (can also specify multiple events by using spaces)
@@ -51,6 +61,7 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	/**
 	 * Attaches an event listener from the specified event emitter.
 	 * If the emitter is a function, it is interpreted as the function to attach events.
+	 * Notice, if you don't provide a group for the operation, it will be taken from the event name.
 	 *
 	 * @param emitter - event emitter
 	 * @param events - event or list of events (can also specify multiple events by using spaces)
@@ -94,14 +105,14 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 
 		const
 			that = this,
-			links: object[] = [],
+			links: IdObject[] = [],
 			multipleEvent = events.length > 1;
 
 		for (let i = 0; i < events.length; i++) {
 			const
 				event = events[i];
 
-			const link = this.registerTask<object>({
+			const link = this.registerTask<IdObject>({
 				...p,
 
 				name: this.namespaces.eventListener,
@@ -159,7 +170,8 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 				}
 			});
 
-			if (link) {
+			// eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+			if (link != null) {
 				links.push(link);
 			}
 		}
@@ -170,6 +182,7 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	/**
 	 * Attaches an event listener from the specified event emitter, but the event is listened only once.
 	 * If the emitter is a function, it is interpreted as the function to attach events.
+	 * Notice, if you don't provide a group for the operation, it will be taken from the event name.
 	 *
 	 * @param emitter - event emitter
 	 * @param events - event or list of events (can also specify multiple events by using spaces)
@@ -186,6 +199,7 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	/**
 	 * Attaches an event listener from the specified event emitter, but the event is listened only once.
 	 * If the emitter is a function, it is interpreted as the function to attach events.
+	 * Notice, if you don't provide a group for the operation, it will be taken from the event name.
 	 *
 	 * @param emitter - event emitter
 	 * @param events - event or list of events (can also specify multiple events with a space)
@@ -225,6 +239,7 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	/**
 	 * Returns a promise that is resolved after emitting the specified event.
 	 * If the emitter is a function, it is interpreted as the function to attach events.
+	 * Notice, if you don't provide a group for the operation, it will be taken from the event name.
 	 *
 	 * @param emitter - event emitter
 	 * @param events - event or list of events (can also specify multiple events with a space)
@@ -241,6 +256,7 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	/**
 	 * Returns a promise that is resolved after emitting the specified event.
 	 * If the emitter is a function, it is interpreted as the function to attach events.
+	 * Notice, if you don't provide a group for the operation, it will be taken from the event name.
 	 *
 	 * @param emitter - event emitter
 	 * @param events - event or list of events (can also specify multiple events with a space)
@@ -296,12 +312,15 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	off(id?: EventId): this;
 
 	/**
-	 * Removes the specified event listener or a group of listeners
+	 * Removes the specified event listener or a group of listeners.
+	 * Notice, you can't remove event listeners by a label without providing a group.
+	 *
 	 * @param opts - options for the operation
 	 */
-	off(opts: ClearOptionsId<EventId>): this;
-	off(task?: EventId | ClearOptionsId<EventId>): this {
-		return this.clearEventListener(task);
+	off(opts: StrictClearOptionsId<EventId>): this;
+
+	off(task?: EventId | StrictClearOptionsId<EventId>): this {
+		return this.clearEventListener(<any>task);
 	}
 
 	/**
@@ -311,11 +330,13 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	clearEventListener(id?: EventId): this;
 
 	/**
-	 * Removes the specified event listener or a group of listeners
+	 * Removes the specified event listener or a group of listeners.
+	 * Notice, you can't remove event listeners by a label without providing a group.
+	 *
 	 * @param opts - options for the operation
 	 */
-	clearEventListener(opts: ClearOptionsId<EventId>): this;
-	clearEventListener(task?: EventId | ClearOptionsId<EventId>): this {
+	clearEventListener(opts: StrictClearOptionsId<EventId>): this;
+	clearEventListener(task?: EventId | StrictClearOptionsId<EventId>): this {
 		if (Object.isArray(task)) {
 			for (let i = 0; i < task.length; i++) {
 				this.clearEventListener(<EventId>task[i]);
@@ -334,12 +355,14 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	muteEventListener(id?: EventId): this;
 
 	/**
-	 * Mutes the specified event listener or a group of listeners
+	 * Mutes the specified event listener or a group of listeners.
+	 * Notice, you can't mute event listeners by a label without providing a group.
+	 *
 	 * @param opts - options for the operation
 	 */
-	muteEventListener(opts: ClearOptionsId<EventId>): this;
-	muteEventListener(task?: EventId | ClearOptionsId<EventId>): this {
-		return this.markEvent('muted', task);
+	muteEventListener(opts: StrictClearOptionsId<EventId>): this;
+	muteEventListener(task?: EventId | StrictClearOptionsId<EventId>): this {
+		return this.markEvent('muted', <any>task);
 	}
 
 	/**
@@ -349,12 +372,14 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	unmuteEventListener(id?: EventId): this;
 
 	/**
-	 * Unmutes the specified event listener or a group of listeners
+	 * Unmutes the specified event listener or a group of listeners.
+	 * Notice, you can't unmute event listeners by a label without providing a group.
+	 *
 	 * @param opts - options for the operation
 	 */
-	unmuteEventListener(opts: ClearOptionsId<EventId>): this;
-	unmuteEventListener(task?: EventId | ClearOptionsId<EventId>): this {
-		return this.markEvent('!muted', task);
+	unmuteEventListener(opts: StrictClearOptionsId<EventId>): this;
+	unmuteEventListener(task?: EventId | StrictClearOptionsId<EventId>): this {
+		return this.markEvent('!muted', <any>task);
 	}
 
 	/**
@@ -364,12 +389,14 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	suspendEventListener(id?: EventId): this;
 
 	/**
-	 * Suspends the specified event listener or a group of listeners
+	 * Suspends the specified event listener or a group of listeners.
+	 * Notice, you can't suspend event listeners by a label without providing a group.
+	 *
 	 * @param opts - options for the operation
 	 */
-	suspendEventListener(opts: ClearOptionsId<EventId>): this;
-	suspendEventListener(task?: EventId | ClearOptionsId<EventId>): this {
-		return this.markEvent('paused', task);
+	suspendEventListener(opts: StrictClearOptionsId<EventId>): this;
+	suspendEventListener(task?: EventId | StrictClearOptionsId<EventId>): this {
+		return this.markEvent('paused', <any>task);
 	}
 
 	/**
@@ -379,12 +406,14 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	unsuspendEventListener(id?: EventId): this;
 
 	/**
-	 * Unsuspends the specified event listener or a group of listeners
+	 * Unsuspends the specified event listener or a group of listeners.
+	 * Notice, you can't unsuspend event listeners by a label without providing a group.
+	 *
 	 * @param opts - options for the operation
 	 */
-	unsuspendEventListener(opts: ClearOptionsId<EventId>): this;
-	unsuspendEventListener(p: EventId | ClearOptionsId<EventId>): this {
-		return this.markEvent('!paused', p);
+	unsuspendEventListener(opts: StrictClearOptionsId<EventId>): this;
+	unsuspendEventListener(p?: EventId | StrictClearOptionsId<EventId>): this {
+		return this.markEvent('!paused', <any>p);
 	}
 
 	/**
@@ -418,8 +447,8 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	 * @param label
 	 * @param opts - additional options
 	 */
-	protected markEvent(label: string, opts: ClearOptionsId<EventId>): this;
-	protected markEvent(label: string, task: EventId | ClearOptionsId<EventId>): this {
+	protected markEvent(label: string, opts: StrictClearOptionsId<EventId>): this;
+	protected markEvent(label: string, task?: EventId | StrictClearOptionsId<EventId>): this {
 		if (Object.isArray(task)) {
 			for (let i = 0; i < task.length; i++) {
 				this.markEvent(label, <EventId>task[i]);