+++ src/core/url/interface.ts
@@ -34,19 +34,26 @@ export interface ToQueryStringOptions {
 	 * @example
 	 * ```js
 	 * // foo[]=1&bar[bla]=2
-	 * fromQueryString({foo: [1], bar: {bla: 2}}, {arraySyntax: true});
+	 * toQueryString({foo: [1], bar: {bla: 2}}, {arraySyntax: true});
 	 * ```
 	 */
 	arraySyntax?: boolean;
 
 	/**
-	 * Returns true if a value must be presented in a query string
+	 * Filters values that shouldn't be serialized.
+	 * By default, the function skip all values with null-s and empty strings.
 	 *
 	 * @param value
-	 * @param key
-	 * @param fullKey - full key for deep objects ({a: {b: 1}} => 'a_b')
+	 * @param key - property key
+	 * @param path - accumulated property path ({a: {b: 1}} => 'a_b')
+	 *
+	 * @example
+	 * ```js
+	 * // foo=1
+	 * toQueryString({foo: 1, bar: {bla: 2}}, {paramsFilter: (el, key) => key !== 'bla'});
+	 * ```
 	 */
-	paramsFilter?(value: unknown, key: string, fullKey?: string): boolean;
+	paramsFilter?(value: unknown, key: string, path?: string): unknown;
 }
 
 export interface FromQueryStringOptions {
@@ -92,13 +99,4 @@ export interface FromQueryStringOptions {
 	 * ```
 	 */
 	arraySyntax?: boolean;
-}
-
-/**
- * Element of stack that used in the `toQueryString` function
- */
-export interface StackItem {
-	key: string;
-	data: unknown;
-	checked?: true;
 }